tilbage/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.5b1
cffi==1.14.5
channels==3.0.3
click==8.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
whitenoise==5.2.0
zope.interface==5.4.0
